move joining life cycle handler to separate file
- move new hud globals outside of header
