+++ xlive/Blam/Engine/interface/user_interface_networking.cpp
@@ -9,6 +9,7 @@
 #include "game/players.h"
 #include "game/player_constants.h"
 #include "main/main_game.h"
+#include "networking/logic/life_cycle/life_cycle_handler_joining.h"
 #include "networking/logic/life_cycle_manager.h"
 #include "networking/logic/network_life_cycle.h"
 #include "networking/logic/network_session_interface.h"
