+++ xlive/Blam/Engine/networking/logic/life_cycle/life_cycle_handler_joining.cpp
@@ -1,0 +1,8 @@
+#include "stdafx.h"
+#include "life_cycle_handler_joining.h"
+
+void __cdecl c_game_life_cycle_handler_joining::check_joining_capability(void)
+{
+	INVOKE(0x1AD643, 0x1A65C0, check_joining_capability);
+	return;
+}
